updated to version 1.3.4

+++ app/src/main/java/com/application/bris/brisi_pemutus/model/rpc/Rpc.java
@@ -4,173 +4,173 @@ import com.google.gson.annotations.SerializedName;
 
 public class Rpc {
     @SerializedName("penghasilanLain")
-    private int penghasilanLain;
+    private Long penghasilanLain;
     @SerializedName("pengeluaranTransportasi")
-    private int pengeluaranTransportasi;
+    private Long pengeluaranTransportasi;
     @SerializedName("rpc")
-    private int rpc;
+    private Long rpc;
     @SerializedName("pengeluaranPajak")
-    private int pengeluaranPajak;
+    private Long pengeluaranPajak;
     @SerializedName("pendapatanUsaha")
-    private int pendapatanUsaha;
+    private Long pendapatanUsaha;
     @SerializedName("rpcRatio")
-    private int rpcRatio;
+    private Float rpcRatio;
     @SerializedName("pendapatanBersih")
-    private int pendapatanBersih;
+    private Long pendapatanBersih;
     @SerializedName("angsuranAplikasi")
-    private int angsuranAplikasi;
+    private Long angsuranAplikasi;
     @SerializedName("pengeluaranHargaSewa")
-    private int pengeluaranHargaSewa;
+    private Long pengeluaranHargaSewa;
     @SerializedName("labaRugi")
-    private int labaRugi;
+    private Long labaRugi;
     @SerializedName("pengeluaranTelpListrik")
-    private int pengeluaranTelpListrik;
+    private Long pengeluaranTelpListrik;
     @SerializedName("pengeluaranUsaha")
-    private int pengeluaranUsaha;
+    private Long pengeluaranUsaha;
     @SerializedName("pengeluaranGajiPegawai")
-    private int pengeluaranGajiPegawai;
+    private Long pengeluaranGajiPegawai;
     @SerializedName("pengeluaranLainnya")
-    private int pengeluaranLainnya;
+    private Long pengeluaranLainnya;
     @SerializedName("pengeluaranRumahTangga")
-    private int pengeluaranRumahTangga;
+    private Long pengeluaranRumahTangga;
     @SerializedName("angsuranBrisExisting")
-    private int angsuranBrisExisting;
+    private Long angsuranBrisExisting;
     @SerializedName("pengeluaranHargaPokokPenjualan")
-    private int pengeluaranHargaPokokPenjualan;
+    private Long pengeluaranHargaPokokPenjualan;
 
-    public int getPenghasilanLain() {
+    public Long getPenghasilanLain() {
         return penghasilanLain;
     }
 
-    public void setPenghasilanLain(int penghasilanLain) {
+    public void setPenghasilanLain(Long penghasilanLain) {
         this.penghasilanLain = penghasilanLain;
     }
 
-    public int getPengeluaranTransportasi() {
+    public Long getPengeluaranTransportasi() {
         return pengeluaranTransportasi;
     }
 
-    public void setPengeluaranTransportasi(int pengeluaranTransportasi) {
+    public void setPengeluaranTransportasi(Long pengeluaranTransportasi) {
         this.pengeluaranTransportasi = pengeluaranTransportasi;
     }
 
-    public int getRpc() {
+    public Long getRpc() {
         return rpc;
     }
 
-    public void setRpc(int rpc) {
+    public void setRpc(Long rpc) {
         this.rpc = rpc;
     }
 
-    public int getPengeluaranPajak() {
+    public Long getPengeluaranPajak() {
         return pengeluaranPajak;
     }
 
-    public void setPengeluaranPajak(int pengeluaranPajak) {
+    public void setPengeluaranPajak(Long pengeluaranPajak) {
         this.pengeluaranPajak = pengeluaranPajak;
     }
 
-    public int getPendapatanUsaha() {
+    public Long getPendapatanUsaha() {
         return pendapatanUsaha;
     }
 
-    public void setPendapatanUsaha(int pendapatanUsaha) {
+    public void setPendapatanUsaha(Long pendapatanUsaha) {
         this.pendapatanUsaha = pendapatanUsaha;
     }
 
-    public int getRpcRatio() {
+    public Float getRpcRatio() {
         return rpcRatio;
     }
 
-    public void setRpcRatio(int rpcRatio) {
+    public void setRpcRatio(Float rpcRatio) {
         this.rpcRatio = rpcRatio;
     }
 
-    public int getPendapatanBersih() {
+    public Long getPendapatanBersih() {
         return pendapatanBersih;
     }
 
-    public void setPendapatanBersih(int pendapatanBersih) {
+    public void setPendapatanBersih(Long pendapatanBersih) {
         this.pendapatanBersih = pendapatanBersih;
     }
 
-    public int getAngsuranAplikasi() {
+    public Long getAngsuranAplikasi() {
         return angsuranAplikasi;
     }
 
-    public void setAngsuranAplikasi(int angsuranAplikasi) {
+    public void setAngsuranAplikasi(Long angsuranAplikasi) {
         this.angsuranAplikasi = angsuranAplikasi;
     }
 
-    public int getPengeluaranHargaSewa() {
+    public Long getPengeluaranHargaSewa() {
         return pengeluaranHargaSewa;
     }
 
-    public void setPengeluaranHargaSewa(int pengeluaranHargaSewa) {
+    public void setPengeluaranHargaSewa(Long pengeluaranHargaSewa) {
         this.pengeluaranHargaSewa = pengeluaranHargaSewa;
     }
 
-    public int getLabaRugi() {
+    public Long getLabaRugi() {
         return labaRugi;
     }
 
-    public void setLabaRugi(int labaRugi) {
+    public void setLabaRugi(Long labaRugi) {
         this.labaRugi = labaRugi;
     }
 
-    public int getPengeluaranTelpListrik() {
+    public Long getPengeluaranTelpListrik() {
         return pengeluaranTelpListrik;
     }
 
-    public void setPengeluaranTelpListrik(int pengeluaranTelpListrik) {
+    public void setPengeluaranTelpListrik(Long pengeluaranTelpListrik) {
         this.pengeluaranTelpListrik = pengeluaranTelpListrik;
     }
 
-    public int getPengeluaranGajiPegawai() {
+    public Long getPengeluaranGajiPegawai() {
         return pengeluaranGajiPegawai;
     }
 
-    public void setPengeluaranGajiPegawai(int pengeluaranGajiPegawai) {
+    public void setPengeluaranGajiPegawai(Long pengeluaranGajiPegawai) {
         this.pengeluaranGajiPegawai = pengeluaranGajiPegawai;
     }
 
-    public int getPengeluaranLainnya() {
+    public Long getPengeluaranLainnya() {
         return pengeluaranLainnya;
     }
 
-    public void setPengeluaranLainnya(int pengeluaranLainnya) {
+    public void setPengeluaranLainnya(Long pengeluaranLainnya) {
         this.pengeluaranLainnya = pengeluaranLainnya;
     }
 
-    public int getPengeluaranRumahTangga() {
+    public Long getPengeluaranRumahTangga() {
         return pengeluaranRumahTangga;
     }
 
-    public void setPengeluaranRumahTangga(int pengeluaranRumahTangga) {
+    public void setPengeluaranRumahTangga(Long pengeluaranRumahTangga) {
         this.pengeluaranRumahTangga = pengeluaranRumahTangga;
     }
 
-    public int getAngsuranBrisExisting() {
+    public Long getAngsuranBrisExisting() {
         return angsuranBrisExisting;
     }
 
-    public void setAngsuranBrisExisting(int angsuranBrisExisting) {
+    public void setAngsuranBrisExisting(Long angsuranBrisExisting) {
         this.angsuranBrisExisting = angsuranBrisExisting;
     }
 
-    public int getPengeluaranHargaPokokPenjualan() {
+    public Long getPengeluaranHargaPokokPenjualan() {
         return pengeluaranHargaPokokPenjualan;
     }
 
-    public void setPengeluaranHargaPokokPenjualan(int pengeluaranHargaPokokPenjualan) {
+    public void setPengeluaranHargaPokokPenjualan(Long pengeluaranHargaPokokPenjualan) {
         this.pengeluaranHargaPokokPenjualan = pengeluaranHargaPokokPenjualan;
     }
 
-    public int getPengeluaranUsaha() {
+    public Long getPengeluaranUsaha() {
         return pengeluaranUsaha;
     }
 
-    public void setPengeluaranUsaha(int pengeluaranUsaha) {
+    public void setPengeluaranUsaha(Long pengeluaranUsaha) {
         this.pengeluaranUsaha = pengeluaranUsaha;
     }
 }
